pyproject.toml
[project]
name = "my-adk-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Zuriel Yahav",email = "zyahav@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
